Create new folder for survey/ move items over
