+++ code/web/src/setup/routes/user.js
@@ -3,7 +3,7 @@ import Login from '../../modules/user/Login'
 import Signup from '../../modules/user/Signup'
 import Profile from '../../modules/user/Profile'
 import Subscriptions from '../../modules/user/Subscriptions'
-import StyleSurvey from '../../modules/user/styleSurvey'
+import StyleSurvey from '../../modules/survey/styleSurvey'
 
 // User routes
 export default {
